+++ R/check_herit_labs.R
@@ -1,10 +1,23 @@
-# used internally by: sim_trait, sim_trait_mvn, cov_trait
-check_labs <- function(
-                       labs,
-                       labs_sigma_sq,
-                       n_ind,
-                       herit
-                       ) {
+# used internally by: sim_trait, cov_trait
+# always run because it checks heritability, and calculates residual variance correctly and consistently!
+check_herit_labs <- function(
+                             herit,
+                             labs,
+                             labs_sigma_sq,
+                             n_ind
+                             ) {
+    # always check herit
+    if ( length( herit ) != 1)
+        stop( '`herit` must be a scalar! (input has length ', length( herit ), ')' )
+    if ( herit < 0 )
+        stop( '`herit` must be non-negative!' )
+    if ( herit > 1 )
+        stop( '`herit` cannot be greater than 1!' )
+    
+    # residual variance when there are no group effects
+    # the previous checks guarantee this is non-negative and smaller or equal than 1
+    sigma_sq_residual <- 1 - herit
+    
     # if there are group effects, check those
     if ( !is.null( labs ) ) {
         if ( is.null( labs_sigma_sq ) )
@@ -24,11 +37,18 @@ check_labs <- function(
             stop( '`labs_sigma_sq` must be non-negative!')
         if ( any( labs_sigma_sq > 1 ) )
             stop( '`labs_sigma_sq` cannot be greater than 1!')
-        if ( sum( labs_sigma_sq ) > 1 - herit )
+        # calculate residual variance correctly for these cases
+        sigma_sq_residual <- 1 - herit - sum( labs_sigma_sq )
+        if ( sigma_sq_residual < 0 )
             stop( 'The sum of `labs_sigma_sq` (', sum( labs_sigma_sq ), ') must not exceed `1 - herit` (', 1 - herit, ')!' )
     }
 
     # `labs` was edited to always be a matrix, so return it!
     # (`labs_sigma_sq` is unchanged)
-    return( labs )
+    return(
+        list(
+            labs = labs,
+            sigma_sq_residual = sigma_sq_residual
+        )
+    )
 }